+++ src/modules/Dropdown/index.tsx
@@ -4,7 +4,7 @@ import {
   ViewStyle,
   ScrollView,
 } from 'react-native'
-import { SlideModal, SlideModalProps, SlideModalState } from '../../components/SlideModal'
+import { SlideModal, SlideModalProps } from '../../components/SlideModal'
 import { Radio } from '../../components/Radio'
 import dropdownStyles from './styles'
 
@@ -23,10 +23,7 @@ export interface DropdownProps extends SlideModalProps {
   onChange: Function
 }
 
-export interface DropdownState extends SlideModalState {
-}
-
-export class Dropdown extends React.Component<DropdownProps, DropdownState> {
+export class Dropdown extends React.Component<DropdownProps> {
   private slideModal = null
 
   static defaultProps = {
@@ -90,7 +87,7 @@ export class Dropdown extends React.Component<DropdownProps, DropdownState> {
       direction === 'down' ? [true, false, false] : [false, false, true]
     )
     return (
-      <SlideModal
+      <SlideModal<SlideModalProps>
         ref={c => {
           this.slideModal = c
         }}