+++ docs/CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+
+## [2.0.3] - 2019-04-23
+### Changed
+- Modal 系列支持 scrollable 属性，在内容溢出屏幕时可以滚动
+- Modal 系列属性 backdropOpacity 去掉，通过 backdropColor 的 rgba 色值控制透明度
+
 ## [2.0.2] - 2019-04-18
 ### Changed
 - Dropdown 实现完善